fixed tests, and separated uninitialized array from constexpr vector
