+++ include/vecl/constexpr_vector.hpp
@@ -2,10 +2,9 @@
 #define VECL_CONSTEXPR_VECTOR_H
 
 #include "config/config.h"
+#include "uninitialized_array.hpp"
 #include <iterator> // advance, distance
 #include <stdexcept> // length_error
-#include <memory> // uninitialized_move, uninitialized_copy
-#include <array> // array
 
 namespace vecl
 {
@@ -29,32 +28,6 @@ namespace vecl
 	template<typename T, size_t N, bool Strict = true>
 	class constexpr_vector
 	{
-		template<typename It>
-		static constexpr void _uninitialized_copy(It s, It e, T* begin)
-		{
-			for (; s != e; ++begin, ++s)
-				std::construct_at(begin, *s);
-		}
-		template<typename It>
-		static constexpr void _uninitialized_move(It s, It e, T* begin)
-		{
-			for (; s != e; ++begin, ++s)
-				std::construct_at(begin, std::move(*s));
-		}
-
-		static constexpr void _uninitialized_fill_n(T* begin, size_t n, const T& value)
-		{
-			for (; n--; ++begin)
-				std::construct_at(begin, value);
-		}
-		static constexpr void _uninitialized_default_construct_n(T* begin, size_t n)
-		{
-			if constexpr (!std::is_trivially_default_constructible_v<T>)
-			{
-				for (; n > 0; ++begin, --n)
-					std::construct_at(begin);
-			}
-		}
 
 		/**
 		 * @return true if ref is in the range.
@@ -1202,61 +1175,6 @@ namespace vecl
 		}
 
 	private:
-
-
-		/**
-		 * @brief array of uninitialized memory
-		 */
-		template<typename T, size_t N>
-		class uninitialized_array
-		{
-		public:
-
-			constexpr uninitialized_array()
-			{
-				if constexpr (_is_sufficiently_trivial)
-					if (std::is_constant_evaluated())
-						for (int i = 0; i < N; ++i)
-							_storage[i] = T();
-			}
-
-			/**
-			 * @brief Direct access to storage
-			 */
-			constexpr const T* data() const VECL_NOEXCEPT
-			{
-				if constexpr (_is_sufficiently_trivial)
-					return static_cast<const T*>(_storage.data());
-				else
-					return reinterpret_cast<const T*>(std::addressof(_storage));
-			}
-
-			/**
-			 * @brief Direct access to storage
-			 */
-			constexpr T* data() VECL_NOEXCEPT
-			{
-				if constexpr (_is_sufficiently_trivial)
-					return static_cast<T*>(_storage.data());
-				else
-					return reinterpret_cast<T*>(std::addressof(_storage));
-			}
-
-		private:
-			static constexpr bool _is_sufficiently_trivial =
-				std::is_trivially_default_constructible_v<T> && std::is_trivially_destructible_v<T>;
-
-			/**
-			 * @deprecated std::aligned_storage_t deprecated in c++23
-			 */
-			using storage_type = std::conditional_t<
-				_is_sufficiently_trivial,
-				std::array<T, N>,
-				std::aligned_storage_t<N * sizeof(T), alignof(T)>
-			>;
-
-			VECL_NO_UNIQUE_ADDRESS storage_type _storage;
-		};
 
 		uninitialized_array<T, N> _buffer;
 		size_type _size = 0;
